feat: Oprs restantes + Requisitos Laysse

+++ lib/src/Column/repositories/ColumnRepository.dart
@@ -1,10 +1,16 @@
 import 'dart:io';
 import 'package:dart_frog/dart_frog.dart';
+import 'package:dart_jsonwebtoken/dart_jsonwebtoken.dart';
+
+import 'package:dotenv/dotenv.dart';
+import 'package:google_cloud_firestore/google_cloud_firestore.dart';
 
 import 'package:hometasks/config/DataBase_client.dart';
 import 'package:hometasks/src/Column/models/ColumnDBModel.dart';
 import 'package:hometasks/src/Column/models/ColumnModel.dart';
 
+///Importação de dados sensíveis
+final _env = DotEnv()..load();
 
 ///Classe para interação direta com o banco remoto
 class ColumnRepository {
@@ -16,20 +22,29 @@ class ColumnRepository {
   //-----------------------------
   ///Operação de criação
   Future<Response> createColumn(
-    ColumnModel column
+    String idTable,
+    ColumnModel column,
+    RequestContext context
     ) async {
-      try{
+      if(await _validateOpr(idTable, 'Editor', context)){
+        try{
 
-        await ref
-        .doc()
-        .set(column.toMap());
-        
+          await ref
+          .doc()
+          .set(column.toMap());
+          
+          return Response.json(
+            statusCode: HttpStatus.created, 
+            body: 'Criação bem sucedida'
+          );
+        }catch(e){
+          throw Exception(e);
+        }
+      }else{
         return Response.json(
-          statusCode: HttpStatus.created, 
-          body: 'Criação bem sucedida'
+          statusCode: HttpStatus.badRequest,
+          body: 'Você não possui acesso à esta operação'
         );
-      }catch(e){
-        throw Exception(e);
       }
   }
 
@@ -38,8 +53,11 @@ class ColumnRepository {
   //-----------------------------
   ///Operação de leitura individual
   Future<Response> readColumn(
-    String id
-      ) async{
+    String idTable,
+    String id,
+    RequestContext context
+    ) async {
+    if(await _validateOpr(idTable, 'Reader', context)){
       try{
         final val = await ref
         .doc(id)
@@ -54,6 +72,45 @@ class ColumnRepository {
       }catch(e){
         throw Exception(e);
       }
+    }else{
+      return Response.json(
+        statusCode: HttpStatus.badRequest,
+        body: 'Você não possui acesso à esta operação'
+      );
+    }
+  }
+
+  //-----------------------------
+  //            read - Reader
+  //-----------------------------
+  ///Operação de leitura em conjunto
+  Future<Response> readAllColumns(
+      String idTable,
+      RequestContext context
+      ) async {
+      if(await _validateOpr(idTable, 'Reader', context)){
+        try{
+          final val = await ref
+          .get();
+          
+          final formDados = <Map<String, dynamic>>[];
+          for (var i = 0; i < val.docs.length; i++){
+            formDados.add(ColumnDBModel.fromFirestore(val.docs[i]).toMap());
+          }
+
+          return Response.json(
+            statusCode: HttpStatus.found, 
+            body: formDados
+          );
+        }catch(e){
+          throw Exception(e);
+        }
+      }else{
+        return Response.json(
+          statusCode: HttpStatus.badRequest,
+          body: 'Você não possui acesso à esta operação'
+        );
+      }
   }
 
   //-----------------------------
@@ -61,20 +118,29 @@ class ColumnRepository {
   //-----------------------------
   ///Operação de atualização individual
   Future<Response> updateColumn(
+    String idTable,
     String id, 
-    ColumnModel column
-    ) async{ 
-      try{
-        await ref
-        .doc(id)
-        .update(column.toMap());
+    ColumnModel column,
+    RequestContext context
+    ) async { 
+      if(await _validateOpr(idTable, 'Editor', context)){
+        try{
+          await ref
+          .doc(id)
+          .update(column.toMap());
 
+          return Response.json(
+            statusCode: HttpStatus.accepted, 
+            body: 'Atualização bem sucedida'
+          );
+        }catch(e){
+          throw Exception(e);
+        }
+      }else{
         return Response.json(
-          statusCode: HttpStatus.accepted, 
-          body: 'Atualização bem sucedida'
+          statusCode: HttpStatus.badRequest,
+          body: 'Você não possui acesso à esta operação'
         );
-      }catch(e){
-        throw Exception(e);
       }
   }
 
@@ -83,19 +149,79 @@ class ColumnRepository {
   //-----------------------------
   ///Operação de remoção individual
   Future<Response> deleteColumn(
-    String id
-    ) async{
-      try{
-        await ref
-        .doc(id)
-        .delete();
+    String idTable,
+    String id,
+    RequestContext context
+    ) async {
+      if(await _validateOpr(idTable, 'Editor', context)){
+        try{
+          await ref
+          .doc(id)
+          .delete();
 
-        return Response(
-          statusCode: HttpStatus.accepted, 
-          body: 'Deleção bem sucedida'
+          return Response(
+            statusCode: HttpStatus.accepted, 
+            body: 'Deleção bem sucedida'
+          );
+        }catch(e){
+          throw Exception(e);
+        }
+      }else{
+        return Response.json(
+          statusCode: HttpStatus.badRequest,
+          body: 'Você não possui acesso à esta operação'
         );
-      }catch(e){
-        throw Exception(e);
       }
+  }
+}
+
+//-----------------------------
+//      Permissão Cargos + RLS
+//-----------------------------
+///Limitar as operações a depender do cargo do usuário
+Future<bool> _validateOpr(
+  String idTable,
+  String cargo,
+  RequestContext context,
+)async{
+  try{
+    //Obtenção de dados do usuário
+    final request = context.request;
+    final header = request.headers['authorization'];
+                
+    final jwt = JWT.verify(
+      header!, 
+      SecretKey(_env['jwtSecretKey'].toString())
+    );
+
+    final payload = jwt.payload as Map<String, dynamic>;
+    
+    //Busca pelo cargo do usuário
+    final relationships = firestore.collection('Relationship');
+
+    final data = await relationships
+      .where(
+        'idUser', 
+        WhereFilter.equal, 
+        payload['id'].toString()
+      )
+      .where(
+        'idTable', 
+        WhereFilter.equal, 
+        idTable
+      )
+      .get();
+
+    if(
+      data.docs.first.data().isNotEmpty
+      &&
+      data.docs.first.data()['cargo'] == cargo
+    ){
+      return true;
+    }else{
+      return false;
+    }
+  }catch(e){
+    return false;
   }
 }
